// src/app/model/job-model.ts
export interface IJobs {
  id: string;
  title: string;
  applyLink: string;
  jobDescription: string;
  companyName: string;
  location: string;
  experience: string;
  salary: string;
  type: string;
  skills: string[];
  startDate: string;
  endDate: string;
  created: string;
  source: string;
  time: number;
}